Write to /etc/hostname_new
Writing directly to /etc/HOSTNAME has an unintended consequence of
actually changing the hostname reported by the "hostname" command. This
breaks everything on X and leaves the user unable to launch any
applications unless they explicitly set the MIT-MAGIC-COOKIE-1 with
xauth add. There is some discussion in this forum thread:
https://forum.salixos.org/viewtopic.php?f=17&t=8335